TeenThrive Teen Wellness/teen_wellness_app_v2_ollama/requirements.txt
ImageHash==4.3.1
Pillow==10.4.0
bcrypt==4.1.2
certifi>=2024.2.2
fastapi==0.111.0
httpx==0.27.0
numpy==1.26.4
pandas==2.2.2
passlib==1.7.4
pydantic==2.8.2
pyjwt==2.8.0
python-multipart==0.0.9
requests>=2.32.3
sqlmodel==0.0.22
streamlit==1.37.0
uvicorn==0.30.1
